Typings
- Separated Types into separate file
